abstract parallel execute
